chore(1-3336): begin moving utils into separate files
